renamed files and changed code to be compliant to the new API
