+++ cogo_widgets.py
@@ -16,8 +16,9 @@ This is a collection of custom QWidgets.
  ***************************************************************************/
 """
 
-from PyQt5.QtCore import Qt
+
 from PyQt5.QtGui import QCursor
+from PyQt5.QtCore import Qt
 from PyQt5.QtWidgets import QPushButton, QDialogButtonBox
